resolved circular dependencies of chat by extracting chat providers into chat_providers and chat data models into chat_message leaf modules
